# contracts/openapi/manoksha-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Manoksha Collections API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/customer/orders/{id}/payment": {
      "get": {
        "tags": [
          "Customer"
        ],
        "operationId": "CustomerOrderPayment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPaymentStatusDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payment-simulator/{providerOrderRef}": {
      "get": {
        "tags": [
          "Payment simulator (development)"
        ],
        "operationId": "SimulatorGetPayment",
        "parameters": [
          {
            "name": "providerOrderRef",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulatorPaymentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payment-simulator/{providerOrderRef}/approve": {
      "post": {
        "tags": [
          "Payment simulator (development)"
        ],
        "operationId": "SimulatorApprovePayment",
        "parameters": [
          {
            "name": "providerOrderRef",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulatorApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulatorPaymentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payment-simulator/{providerOrderRef}/decline": {
      "post": {
        "tags": [
          "Payment simulator (development)"
        ],
        "operationId": "SimulatorDeclinePayment",
        "parameters": [
          {
            "name": "providerOrderRef",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulatorDeclineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulatorPaymentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/payments": {
      "get": {
        "tags": [
          "Payments"
        ],
        "operationId": "ListPaymentAttempts",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "purpose",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "referenceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentAttemptDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/payment-reconciliations": {
      "get": {
        "tags": [
          "Payments"
        ],
        "operationId": "ListPaymentReconciliations",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentReconciliationDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/payment-reconciliations/{id}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "operationId": "GetPaymentReconciliation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentReconciliationDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OnlinePaymentDto": {
        "type": "object",
        "properties": {
          "attemptId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "redirectUrl": {
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OrderPaymentStatusDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string",
            "nullable": true
          },
          "orderStatus": {
            "type": "string",
            "nullable": true
          },
          "payment": {
            "$ref": "#/components/schemas/OnlinePaymentDto"
          }
        },
        "additionalProperties": false
      },
...
      "PaymentAttemptDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "purpose": {
            "type": "string",
            "nullable": true
          },
          "referenceId": {
            "type": "string",
            "format": "uuid"
          },
          "referenceNumber": {
            "type": "string",
            "nullable": true
          },
          "payerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "providerOrderRef": {
            "type": "string",
            "nullable": true
          },
          "providerPaymentRef": {
            "type": "string",
            "nullable": true
          },
          "initiatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "failureReason": {
            "type": "string",
            "nullable": true
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentStatusChangeDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PaymentReconciliationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "caseNumber": {
            "type": "string",
            "nullable": true
          },
          "paymentAttemptId": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "nullable": true
          },
          "providerOrderRef": {
            "type": "string",
            "nullable": true
          },
          "providerPaymentRef": {
            "type": "string",
            "nullable": true
          },
          "expectedAmount": {
            "type": "number",
            "format": "double"
          },
          "paidAmount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "payerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "purpose": {
            "type": "string",
            "nullable": true
          },
          "referenceId": {
            "type": "string",
            "format": "uuid"
          },
          "referenceNumber": {
            "type": "string",
            "nullable": true
          },
          "reasonCode": {
            "type": "string",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "ownerAction": {
            "type": "string",
            "nullable": true
          },
          "externalRefundRef": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PaymentStatusChangeDto": {
        "type": "object",
        "properties": {
          "fromStatus": {
            "type": "string",
            "nullable": true
          },
          "toStatus": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "SimulatorApproveRequest": {
        "type": "object",
        "properties": {
          "paidAmount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "sendWebhook": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "SimulatorDeclineRequest": {
        "type": "object",
        "properties": {
          "sendWebhook": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "SimulatorPaymentDto": {
        "type": "object",
        "properties": {
          "providerOrderRef": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "windowEnded": {
            "type": "boolean"
          },
          "paidAmount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "providerPaymentRef": {
            "type": "string",
            "nullable": true
          },
          "returnPath": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}